Wizard's Adventure Game (My First Full Unity Game)
In this game you play as a wizard, you are pretty powerful and also have a pretty powerful apatite! This game takes you on an adventure from a dark dungeon, to a lush forest, to a brisk and beautiful mountain top. This game is brimming with platforming and plenty of magic missiles, but the best part of any story is the people you meet along the way!
I spent the entire summer of 2019 learning Unity by making this game and improving upon it. Most of the assets here I made by hand, however! the Forest tileset and the Mountain tileset are both from the amazing

Here is a link to the game complete (it is free to download)
https://casem.itch.io/wizards-adventure